Rename js -> ts
